Fix parameter default value editing
Parameter name key is "name", not "parameter_name" for default values
that come from parameter definition DB items.

Re #2655
