+++ tests/spine_db_editor/widgets/helpers.py
@@ -10,18 +10,15 @@
 # this program. If not, see <http://www.gnu.org/licenses/>.
 ######################################################################################################################
 
-"""Helper utilites for unit tests that test Database manager's table and tree views."""
+"""Helper utilities for unit tests that test Database editor's table and tree views."""
 from types import MethodType
-import unittest
 from unittest import mock
 from PySide6.QtCore import QEvent, Qt
 from PySide6.QtGui import QKeyEvent
 from PySide6.QtWidgets import QApplication
-from spinetoolbox.spine_db_editor.widgets.spine_db_editor import SpineDBEditor
 from spinetoolbox.spine_db_editor.widgets.add_items_dialogs import AddEntityClassesDialog, AddEntitiesDialog
 from spinetoolbox.helpers import signal_waiter
 from spinetoolbox.spine_db_editor.widgets.custom_editors import SearchBarEditor
-from tests.mock_helpers import TestSpineDBManager
 
 
 class EditorDelegateMocking:
@@ -118,42 +115,3 @@ def add_entity(view, name, entity_class_index=0):
     class_index = model.index(entity_class_index, 0, root_index)
     view._context_item = model.item_from_index(class_index)
     add_entity_tree_item({0: name}, view, "Add entities", AddEntitiesDialog)
-
-
-class TestBase(unittest.TestCase):
-    """Base class for Database editor's table and tree view tests."""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.db_codename = cls.__name__ + "_db"
-        if not QApplication.instance():
-            QApplication()
-
-    def _common_setup(self, url, create):
-        with mock.patch("spinetoolbox.spine_db_editor.widgets.spine_db_editor.SpineDBEditor.restore_ui"), mock.patch(
-            "spinetoolbox.spine_db_editor.widgets.spine_db_editor.SpineDBEditor.show"
-        ):
-            mock_settings = mock.MagicMock()
-            mock_settings.value.side_effect = lambda *args, **kwargs: 0
-            self._db_mngr = TestSpineDBManager(mock_settings, None)
-            logger = mock.MagicMock()
-            self._db_map = self._db_mngr.get_db_map(url, logger, codename=self.db_codename, create=create)
-            self._db_editor = SpineDBEditor(self._db_mngr, {url: self.db_codename})
-        QApplication.processEvents()
-
-    def _common_tear_down(self):
-        with mock.patch(
-            "spinetoolbox.spine_db_editor.widgets.spine_db_editor.SpineDBEditor.save_window_state"
-        ), mock.patch("spinetoolbox.spine_db_manager.QMessageBox"):
-            self._db_editor.close()
-        self._db_mngr.close_all_sessions()
-        while not self._db_map.closed:
-            QApplication.processEvents()
-        self._db_mngr.clean_up()
-        self._db_editor.deleteLater()
-        self._db_editor = None
-
-    def _commit_changes_to_database(self, commit_message):
-        with mock.patch.object(self._db_editor, "_get_commit_msg") as commit_msg:
-            commit_msg.return_value = commit_message
-            self._db_editor.ui.actionCommit.trigger()
