+++ tests/spine_db_editor/widgets/test_tabular_view_mixin.py
@@ -18,7 +18,7 @@ from PySide6.QtCore import QItemSelectionModel
 from PySide6.QtWidgets import QApplication
 from spinedb_api import Map
 from tests.mock_helpers import fetch_model
-from tests.spine_db_editor.widgets.helpers import TestBase
+from tests.spine_db_editor.helpers import TestBase
 
 
 class TestPivotHeaderDraggingAndDropping(TestBase):
